+++ src/dts/Serailizers/FugueTree/State.ts
@@ -1,16 +1,22 @@
-import { compress, decompress } from "lz4js";
 import { FTree } from "../../FugueTree/index.js";
+import { COMP } from "../General.js";
+import Pako from "pako";
 
 function serialize(state: FTree): Uint8Array<ArrayBufferLike> {
     // return compress(state.save())
-    return state.save();
+    let bytes = state.save();
+    if (COMP) {
+        bytes = Pako.gzip(bytes);
+    }
+    return bytes;
 }
 
-function deserialize(compressed: Uint8Array<ArrayBufferLike>): FTree {
-    // const raw = decompress(compressed);
-    const raw = compressed;
+function deserialize(bytes: Uint8Array<ArrayBufferLike>): FTree {
+    if (COMP) {
+        bytes = Pako.ungzip(bytes);
+    }
     const state = new FTree();
-    state.load(raw as Uint8Array);
+    state.load(bytes);
     return state;
 }